refactor: EvidenceSection component
