+++ src/components/Hypothesis/EvidenceSection/index.tsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from "react";
 import dayjs from "dayjs";
 import { ChevronRightIcon, DotsVerticalIcon } from "@heroicons/react/outline";
 import { BsPlusLg } from "react-icons/bs";
-import { LogsTable } from "../Logs/Table/logs-table";
-import { TopologyCard } from "../Topology";
-import { Icon } from "../Icon";
+import { LogsTable } from "../../Logs/Table/logs-table";
+import { TopologyCard } from "../../Topology";
+import { Icon } from "../../Icon";
 
 export function EvidenceItem({ evidence }) {
   if (evidence.type === "log") {
